+++ src/sys/ags_gm3.cpp
@@ -128,8 +128,8 @@ void AGS::load_gm3(uint8* data, int page, int transparent)
 				c[6] = ((b0 >> 1) & 1) | ((b1     ) & 2) | ((b2 << 1) & 4);
 				c[7] = ((b0     ) & 1) | ((b1 << 1) & 2) | ((b2 << 2) & 4);
 
-				uint32* dest0 = &vram[dest_screen][(y + sy) * 2 + 0][(x + sx) * 8];
-				uint32* dest1 = &vram[dest_screen][(y + sy) * 2 + 1][(x + sx) * 8];
+				uint8_t* dest0 = &vram[dest_screen][(y + sy) * 2 + 0][(x + sx) * 8];
+				uint8_t* dest1 = &vram[dest_screen][(y + sy) * 2 + 1][(x + sx) * 8];
 				if(transparent == -1) {
 					for(int i = 0; i < 8; i++) {
 						dest0[i] = dest1[i] = c[i] | base;
